Teleporter is back. AI doesn't know how to use it yet.

+++ src/data/entity.h
@@ -9,7 +9,7 @@ namespace VI
 
 typedef u8 Family;
 typedef u16 Revision;
-typedef u32 ComponentMask;
+typedef u64 ComponentMask;
 
 const Family MAX_FAMILIES = sizeof(ComponentMask) * 8;
 #define MAX_ENTITIES 2048
